server/requirements.txt
websockets==15.0.1
aiohttp==3.9.1
pyautogui==0.9.53
segno==1.6.1
python-dotenv==1.0.1

# Screen streaming (WebRTC)
aiortc==1.9.0
mss==9.0.1
numpy>=1.24.0
Pillow>=10.0.0
av>=12.0.0
